requirements.txt
click==8.1.3
colorama==0.4.5
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
platformdirs==2.5.2
protobuf==3.20.1
virtualenv==20.16.5
Werkzeug==2.2.2
